+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - 增加路由请求事件转发功能 forward
 - 移除 Event::dump 方法
 - 集成 Swoole 服务组件
+- 增加 Twig 扩展, `FastD\Framework\Extensions\TplExtension`
 
 
 ### v1.0.* (07/09/2015)